requirements.txt
# 为什么这里保留的是“项目运行所需的核心依赖”：
# 当前项目还是教学级骨架，暂时不加入数据库、测试框架、向量库等扩展依赖，
# 这样可以让安装更快、结构更清楚，也更适合从 0 开始学习。
#
# 为什么不在这一阶段塞入很多“未来可能会用到”的包：
# 因为依赖越多，理解成本越高，排错成本也越高。
# 对当前目标来说，只保留真正会用到的运行依赖，是更简单、更稳妥的方案。

# FastAPI：
# 负责提供 Web API 能力，是当前项目的后端服务框架。
fastapi>=0.115.0,<1.0.0

# Uvicorn：
# 负责运行 FastAPI 应用，是常见的 ASGI 服务启动器。
# `[standard]` 会额外安装一些常用性能/体验增强依赖。
uvicorn[standard]>=0.30.0,<1.0.0

# python-dotenv：
# 负责把 `.env` 文件中的配置加载到环境变量中。
python-dotenv>=1.0.1,<2.0.0

# pydantic：
# 负责数据校验和结构化定义。
# 当前在请求体、响应体和配置模型中都会使用它。
pydantic>=2.8.0,<3.0.0

# pydantic-settings：
# 负责更规范地管理应用配置，和 FastAPI / Pydantic 生态配合很好。
pydantic-settings>=2.3.0,<3.0.0

# httpx：
# 负责发送异步 HTTP 请求。
# 当前项目已经用它来：
# 1. 异步调用 LLM
# 2. 异步调用 embedding
# 3. 执行并发测试脚本
# 它也是这次并发优化的关键依赖之一。
httpx>=0.27.0,<1.0.0

# redis：
# 这是本次 Memory 系统和 Cache 系统优先使用的存储客户端。
# 如果本地没有 Redis 服务，代码会自动退回到内存版实现，
# 但安装这个包后，项目就具备了直接接 Redis 的能力。
redis>=5.0.0,<6.0.0

# python-multipart：
# FastAPI 处理表单和文件上传时会依赖这个包。
# 这次为了支持真正的文件上传入库，需要把它补进来。
python-multipart>=0.0.9,<1.0.0

# sentence-transformers：
# 用于在本地加载 `all-MiniLM-L6-v2` embedding 模型。
# 这样 RAG 在构建向量和查询向量时，就不再依赖远程 OpenAI embedding 接口，
# 安装后即可在本地直接运行：
# `pip install sentence-transformers`
sentence-transformers>=3.0.0,<4.0.0

# LangChain 生态依赖：
# 这几项依赖用于新增“工业化封装版”的 RAG 实现。
# 其中：
# - `langchain` 提供核心抽象
# - `langchain-community` 提供 FAISS 等常用社区组件
# - `langchain-huggingface` 提供本地 HuggingFace embedding 封装
# - `faiss-cpu` 提供本地向量索引能力
langchain>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
langchain-huggingface>=0.1.0,<0.2.0
faiss-cpu>=1.8.0,<2.0.0
